test: split monolithic DatabaseSeeds.swift into per-version files
The 1020-line monolithic fixture tripped SwiftLint errors
(file_length >1000, type_body_length >800, and 13 INSERT lines
>200 chars). Split into one file per frozen version under
Fixtures/Seeds/ and reformatted the offending INSERTs multi-line.
No test-logic or golden-shape changes.
